Update Member tasks
Removed add/remove member tasks and updated member search to view member
