--- src/components/AddMember/AddMember.css
@@ -1,253 +1,0 @@
-.add-member-container {
-  padding: 1.5rem;
-  width: 100%;
-}
-
-.page-title {
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-  text-align: center;
-}
-
-.error-message {
-  background-color: #fed7d7;
-  color: #c53030;
-  padding: 1rem;
-  border-radius: 8px;
-  margin-bottom: 1.5rem;
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  border-left: 4px solid #c53030;
-}
-
-.add-member-form {
-  max-width: 600px;
-  margin: 0 auto;
-  background-color: white;
-  padding: 2rem;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.form-grid {
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  gap: 2rem;
-  margin-bottom: 2rem;
-}
-
-.member-details-section {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.form-group {
-  margin-bottom: 1.5rem;
-}
-
-.form-group label {
-  display: block;
-  margin-bottom: 0.5rem;
-  font-weight: 600;
-  color: #2d3748;
-}
-
-.form-group textarea {
-  resize: vertical;
-  min-height: 100px;
-}
-
-.form-group input,
-.form-group textarea {
-  width: 100%;
-  padding: 0.75rem;
-  border: 1px solid #cbd5e0;
-  border-radius: 4px;
-  font-size: 1rem;
-}
-
-.form-group input:focus,
-.form-group textarea:focus {
-  border-color: #1a365d;
-  box-shadow: 0 0 0 3px rgba(26, 54, 93, 0.2);
-  outline: none;
-}
-
-.form-actions {
-  display: flex;
-  justify-content: center;
-  margin-top: 2rem;
-}
-
-.add-button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  padding: 0.75rem 2rem;
-  background-color: #1a365d;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s, transform 0.2s;
-  text-transform: uppercase;
-  font-weight: 600;
-  min-width: 150px;
-}
-
-.add-button:hover:not(:disabled) {
-  background-color: #2a4365;
-  transform: translateY(-2px);
-}
-
-.add-button:disabled {
-  background-color: #a0aec0;
-  cursor: not-allowed;
-}
-
-/* Confirmation screen styles */
-.confirmation-container {
-  background-color: white;
-  padding: 2.5rem;
-  border-radius: 12px;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
-  max-width: 500px;
-  margin: 2rem auto;
-  text-align: center;
-}
-
-.confirmation-icon {
-  font-size: 3.5rem;
-  color: #48bb78;
-  margin-bottom: 1.5rem;
-  background-color: rgba(72, 187, 120, 0.1);
-  width: 80px;
-  height: 80px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto 1.5rem;
-}
-
-.confirmation-container h2 {
-  color: #2d3748;
-  margin-bottom: 1.5rem;
-  font-size: 1.8rem;
-}
-
-.confirmation-container p {
-  text-align: left;
-  margin-bottom: 2rem;
-  line-height: 1.8;
-  background-color: #f7fafc;
-  padding: 1.5rem;
-  border-radius: 8px;
-}
-
-.confirmation-buttons {
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-}
-
-.done-button,
-.search-button {
-  padding: 0.75rem 1.5rem;
-  border: none;
-  border-radius: 4px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: all 0.3s;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.done-button {
-  background-color: #1a365d;
-  color: white;
-}
-
-.done-button:hover {
-  background-color: #2a4365;
-  transform: translateY(-2px);
-}
-
-.search-button {
-  background-color: #4a5568;
-  color: white;
-}
-
-.search-button:hover {
-  background-color: #2d3748;
-  transform: translateY(-2px);
-}
-
-/* Responsive styles */
-@media (max-width: 768px) {
-  .form-grid {
-    grid-template-columns: 1fr;
-    gap: 1.5rem;
-  }
-
-  .cover-upload-area {
-    max-width: 250px;
-    margin: 0 auto;
-  }
-
-  .form-row {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-
-  .form-group.full-width {
-    grid-column: span 1;
-  }
-
-  .page-title {
-    font-size: 1.8rem;
-    margin-bottom: 1.5rem;
-  }
-
-  .confirmation-container {
-    padding: 1.5rem;
-  }
-
-  .confirmation-buttons {
-    flex-direction: column;
-  }
-}
-
-.small-text {
-  font-size: 0.8rem;
-  color: #718096;
-  margin-top: 0.5rem;
-}
-
-.button-icon {
-  font-size: 0.9rem;
-}
-
-.spinner {
-  display: inline-block;
-  width: 1rem;
-  height: 1rem;
-  border: 2px solid rgba(255, 255, 255, 0.3);
-  border-radius: 50%;
-  border-top-color: white;
-  animation: spin 1s ease-in-out infinite;
-  margin-right: 0.5rem;
-}
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
