--- src/components/MemberSearch/MemberSearch.css
@@ -1,89 +1,0 @@
-.member-search-container {
-  padding: 1.5rem;
-  width: 100%;
-}
-
-.page-title {
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-  text-align: center;
-}
-
-.search-container {
-  display: flex;
-  gap: 1rem;
-  margin-bottom: 2rem;
-  max-width: 600px;
-  margin: 0 auto 2rem;
-}
-
-.search-input {
-  flex: 1;
-  padding: 0.75rem;
-  border: 1px solid #cbd5e0;
-  border-radius: 4px;
-  font-size: 1rem;
-}
-
-.search-button {
-  padding: 0.75rem 1.5rem;
-  background-color: #1a365d;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.search-button:hover {
-  background-color: #2a4365;
-}
-
-.results-placeholder {
-  background-color: #f7fafc;
-  padding: 2rem;
-  border-radius: 8px;
-  text-align: center;
-  color: #718096;
-  margin-top: 2rem;
-}
-
-.search-results-container {
-  max-height: 400px; /* Set a fixed height for the slider */
-  overflow-y: auto; /* Enables vertical scrolling */
-  margin-top: 2rem;
-  border: 1px solid #ddd; /* Optional border for the slider container */
-  border-radius: 8px; /* Rounded corners */
-}
-
-.search-results-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.search-results-table th, 
-.search-results-table td {
-  padding: 12px 15px;
-  text-align: left;
-  border: 1px solid #ddd;
-}
-
-.search-results-table th {
-  background-color: #f2f2f2;
-  font-weight: bold;
-}
-
-.search-results-table tr:hover {
-  background-color: #e2e8f0;
-}
-
-@media (max-width: 768px) {
-  .search-container {
-    flex-direction: column;
-  }
-
-  .page-title {
-    font-size: 1.8rem;
-  }
-}
