--- src/components/RemoveMember/RemoveMember.css
@@ -1,292 +1,0 @@
-.remove-member-container {
-  padding: 1.5rem;
-  width: 100%;
-}
-
-.page-title {
-  font-size: 2.5rem;
-  margin-bottom: 2rem;
-  text-align: center;
-}
-
-.search-container {
-  display: flex;
-  gap: 1rem;
-  margin-bottom: 2rem;
-  max-width: 600px;
-  margin: 0 auto 2rem;
-}
-
-.search-input {
-  flex: 1;
-  padding: 0.75rem;
-  border: 1px solid #cbd5e0;
-  border-radius: 4px;
-  font-size: 1rem;
-}
-
-.search-button {
-  padding: 0.75rem 1.5rem;
-  background-color: #1a365d;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.search-button:hover {
-  background-color: #2a4365;
-}
-
-.results-placeholder {
-  background-color: #f7fafc;
-  padding: 2rem;
-  border-radius: 8px;
-  text-align: center;
-  color: #718096;
-  margin-top: 2rem;
-}
-
-.search-results-container {
-  max-height: 400px; /* Set a fixed height for the slider */
-  overflow-y: auto; /* Enables vertical scrolling */
-  margin-top: 2rem;
-  border: 1px solid #ddd; /* Optional border for the slider container */
-  border-radius: 8px; /* Rounded corners */
-}
-
-.search-results-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.search-results-table th, 
-.search-results-table td {
-  padding: 12px 15px;
-  text-align: left;
-  border: 1px solid #ddd;
-}
-
-.search-results-table th {
-  background-color: #f2f2f2;
-  font-weight: bold;
-}
-
-.search-results-table tr:hover {
-  background-color: #e2e8f0;
-}
-
-.confirmation-dialog {
-  max-width: 500px;
-  margin: 2rem auto 0;
-  background-color: white;
-  padding: 2rem;
-  border-radius: 8px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  border-left: 4px solid #e53e3e;
-}
-
-.confirmation-dialog h2 {
-  color: #e53e3e;
-  margin-bottom: 1rem;
-}
-
-.warning-text {
-  color: #e53e3e;
-  font-weight: 600;
-  margin-top: 0.5rem;
-}
-
-.action-buttons {
-  display: flex;
-  justify-content: flex-end;
-  gap: 1rem;
-  margin-top: 1.5rem;
-}
-
-.cancel-button,
-.delete-confirm-button {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1.25rem;
-  border: none;
-  border-radius: 4px;
-  font-size: 0.95rem;
-  cursor: pointer;
-  transition: all 0.3s;
-  white-space: nowrap;
-}
-
-.cancel-button {
-  background-color: #e2e8f0;
-  color: #4a5568;
-}
-
-.cancel-button:hover {
-  background-color: #cbd5e0;
-}
-
-.delete-button,
-.delete-confirm-button {
-  background-color: #e53e3e;
-  color: white;
-}
-
-.delete-button:hover,
-.delete-confirm-button:hover {
-  background-color: #c53030;
-}
-
-/* Modal styles */
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.delete-modal {
-  background-color: white;
-  border-radius: 8px;
-  padding: 2rem;
-  width: 90%;
-  max-width: 800px;
-  max-height: 90vh;
-  overflow-y: auto;
-  overflow-x: hidden;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-  box-sizing: border-box;
-}
-
-.delete-modal h2 {
-  margin-bottom: 1.5rem;
-  color: #1a365d;
-  font-size: 1.5rem;
-  border-bottom: 2px solid #e2e8f0;
-  padding-bottom: 0.75rem;
-}
-
-.modal-actions {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 1.5rem;
-  flex-wrap: nowrap;
-}
-
-.cancel-button:hover {
-  background-color: #cbd5e0;
-}
-
-.remove-button {
-  padding: 0.75rem 1.5rem;
-  background-color: #e53e3e;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.remove-button:hover {
-  background-color: #c53030;
-}
-
-@media (max-width: 768px) {
-  .search-container {
-    flex-direction: column;
-  }
-
-  .page-title {
-    font-size: 1.8rem;
-  }
-
-  .confirmation-dialog {
-    padding: 1.5rem;
-  }
-
-  .action-buttons {
-    flex-direction: column;
-  }
-}
-
-/* Success Confirmation Dialog */
-.success-confirmation-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 2000;
-}
-
-.success-confirmation {
-  background-color: white;
-  border-radius: 8px;
-  padding: 2.5rem 2rem;
-  width: 90%;
-  max-width: 400px;
-  text-align: center;
-  position: relative;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-.close-btn {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  cursor: pointer;
-  font-size: 1.2rem;
-  color: #718096;
-}
-
-.success-confirmation h2 {
-  font-size: 1.8rem;
-  margin-bottom: 0.5rem;
-  color: #1a365d;
-}
-
-.success-confirmation h3 {
-  font-size: 1.5rem;
-  margin-bottom: 0.5rem;
-  font-style: italic;
-  color: #2d3748;
-}
-
-.success-confirmation p {
-  font-size: 1.3rem;
-  margin-bottom: 1.5rem;
-  color: #4a5568;
-}
-
-.confirmation-icon {
-  width: 80px;
-  height: 80px;
-  border-radius: 50%;
-  border: 3px solid #38a169;
-  margin: 0 auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.checkmark {
-  width: 30px;
-  height: 50px;
-  border-bottom: 3px solid #38a169;
-  border-right: 3px solid #38a169;
-  transform: rotate(45deg);
-  margin-top: -10px;
-}
