+++ src/components/ViewMember/ViewMember.css
@@ -1,0 +1,232 @@
+.member-search-container {
+  padding: 1.5rem;
+  width: 100%;
+}
+
+.page-title {
+  font-size: 2.5rem;
+  margin-bottom: 2rem;
+  text-align: center;
+}
+
+.search-container {
+  display: flex;
+  gap: 1rem;
+  margin-bottom: 2rem;
+  max-width: 600px;
+  margin: 0 auto 2rem;
+}
+
+.search-input {
+  flex: 1;
+  padding: 0.75rem;
+  border: 1px solid #cbd5e0;
+  border-radius: 4px;
+  font-size: 1rem;
+}
+
+.search-button {
+  padding: 0.75rem 1.5rem;
+  background-color: #1a365d;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+.search-button:hover {
+  background-color: #2a4365;
+}
+
+.results-placeholder {
+  background-color: #f7fafc;
+  padding: 2rem;
+  border-radius: 8px;
+  text-align: center;
+  color: #718096;
+  margin-top: 2rem;
+}
+
+.search-results-container {
+  max-height: 400px; /* Set a fixed height for the slider */
+  overflow-y: auto; /* Enables vertical scrolling */
+  margin-top: 2rem;
+  border: 1px solid #ddd; /* Optional border for the slider container */
+  border-radius: 8px; /* Rounded corners */
+}
+
+.search-results-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.search-results-table th, 
+.search-results-table td {
+  padding: 12px 15px;
+  text-align: left;
+  border: 1px solid #ddd;
+}
+
+.search-results-table th {
+  background-color: #f2f2f2;
+  font-weight: bold;
+}
+
+.search-results-table tr:hover {
+  background-color: #e2e8f0;
+}
+
+/* Modal styles */
+.member-details-modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+  padding: 1rem;
+  backdrop-filter: blur(5px);
+}
+
+/* container */
+.member-details {
+  display:flex;
+}
+
+/* columns */
+.member-details > * {
+  width:calc(100% / 3);
+  padding:1rem;
+}
+
+
+.modal-content {
+  background-color: white;
+  border-radius: 12px;
+  padding: 2.5rem;
+  width: 90%;
+  max-width: 900px;
+  max-height: 90vh;
+  overflow-y: auto;
+  box-shadow: 0 15px 30px rgba(0, 0, 0, 0.2);
+  position: relative;
+  display: flex;
+  flex-direction: column;
+}
+
+.modal-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 2rem;
+  border-bottom: 1px solid #e2e8f0;
+  padding-bottom: 1.5rem;
+  width: 100%;
+}
+
+.modal-header h2 {
+  margin: 0;
+  color: #1a365d;
+  font-size: 2rem;
+  flex: 1;
+  font-weight: 700;
+}
+
+.member-details-column h2 {
+  margin: 0;
+  color: #1a365d;
+  font-size: 1.5rem;
+  flex: 1;
+  font-weight: 700;
+}
+
+.loans-column h2 {
+  margin: 0;
+  color: #1a365d;
+  font-size: 1.5rem;
+  flex: 1;
+  font-weight: 700;
+}
+
+.book-cover {
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+}
+
+.book-detail-image {
+  max-width: 100%;
+  max-height: 200px;
+  object-fit: contain;
+  border-radius: 8px;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+.book-details {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.detail-item {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.detail-label {
+  font-weight: 600;
+  color: #718096;
+  font-size: 0.95rem;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
+
+.detail-value {
+  font-size: 1.1rem;
+  color: #2d3748;
+}
+
+.close-modal-button {
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  color: #718096;
+  cursor: pointer;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 0.2s;
+  margin-left: 1rem;
+  border-radius: 50%;
+}
+
+.close-modal-button:hover {
+  color: #1a365d;
+  background-color: #f7fafc;
+}
+
+.modal-content {
+  padding: 1.5rem;
+}
+
+.modal-header h2 {
+  font-size: 1.5rem;
+}
+
+@media (max-width: 768px) {
+  .search-container {
+    flex-direction: column;
+  }
+
+  .page-title {
+    font-size: 1.8rem;
+  }
+}
