+++ src/services/MemberService.js
@@ -22,27 +22,4 @@ const loadMembers = async () => {
 
 // Get all members
 export const getAllMembers = async () => {
-  return await loadMembers();
-};
-
-// Add a new member
-export const addMember = async (newMember) => {
-  const members = await loadMembers();
-
-  // Generate a new ID
-  const maxId =
-    members.length > 0 ? Math.max(...members.map((member) => member.id)) : 0;
-  const memberWithId = {
-    ...newMember,
-    id: maxId + 1,
-  };
-
-  // Add to our in-memory data
-  members.push(memberWithId);
-  membersData = members;
-
-  // In a real app, this would be an API call to save the member
-  console.log("Member added:", memberWithId);
-
-  return memberWithId;
-};
+  return await loadMembers();}
